chore: fix NTI relationship-endpoint only returning uids DHIS2-12625 (#9841) (#9850)
This reverts commit b4ebacfa58418a688e29b75bc3ef26b8fb55cd36.

Reason being that any entity exported from NTI needs to be importable
without modification. Therefore, import/export models need to align.
