+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/domain/mapper/DomainMapper.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.webapi.controller.tracker.export;
+package org.hisp.dhis.tracker.domain.mapper;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -33,11 +33,11 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-public interface DomainMapper<F, T>
+public interface DomainMapper<FROM, TO>
 {
-    T from( F from );
+    TO from( FROM from );
 
-    default List<T> fromCollection( Collection<F> froms )
+    default List<TO> fromCollection( Collection<FROM> froms )
     {
         return Optional.ofNullable( froms )
             .orElse( Collections.emptySet() )
